renaissance-codebase-v1-working/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0

# OCR & Vision
PyMuPDF>=1.23.0
Pillow>=10.0.0
opencv-python>=4.8.0

# LLM/VLM
peft>=0.10.0
qwen-vl-utils>=0.0.2

# Training
datasets>=2.18.0
evaluate>=0.4.0

# Evaluation
jiwer>=3.0.0
editdistance>=0.8.1

# Document parsing
python-docx>=1.1.0

# Utilities
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
scipy>=1.10.0
